pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cogs-archive"
version = "0.1.0"
description = "COGS Data Archive tooling for Zenodo-backed datasets"
requires-python = ">=3.10"

dependencies = [
    "requests",
    "pyyaml",
]

[project.optional-dependencies]
test = [
    "pytest",
    "pytest-mock",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
